+++ redmine/table_redmine_tracker.go
@@ -7,6 +7,7 @@ import (
 	rm "github.com/nixys/nxs-go-redmine/v5"
 	"github.com/turbot/steampipe-plugin-sdk/v5/grpc/proto"
 	"github.com/turbot/steampipe-plugin-sdk/v5/plugin"
+	"github.com/turbot/steampipe-plugin-sdk/v5/plugin/transform"
 )
 
 //// TABLE DEFINITION
@@ -33,11 +34,13 @@ func tableRedmineTracker() *plugin.Table {
 		},
 		Columns: []*plugin.Column{
 			{Name: "id", Type: proto.ColumnType_INT, Description: "The tracker ID."},
-			{Name: "name", Type: proto.ColumnType_STRING, Description: "The tracker name."},
 			{Name: "default_status_id", Type: proto.ColumnType_INT, Description: "Default status ID for new issues."},
 			{Name: "default_status_name", Type: proto.ColumnType_STRING, Description: "Default status name for new issues."},
 			{Name: "description", Type: proto.ColumnType_STRING, Description: "The tracker description."},
 			{Name: "enabled_standard_fields", Type: proto.ColumnType_JSON, Description: "Standard fields enabled for this tracker."},
+			{Name: "name", Type: proto.ColumnType_STRING, Description: "The tracker name."},
+			// Standard columns
+			{Name: "title", Type: proto.ColumnType_STRING, Description: "The display name for this resource.", Transform: transform.FromField("Name")},
 		},
 	}
 }